requirements.txt
flask
flask-cors
livekit-plugins-google>=0.1.0
sentence-transformers         
faiss-cpu        
livekit>=0.19.1
livekit-agents>=0.12.11,<1.0.0
livekit-plugins-openai>=0.10.17,<1.0.0
python-dotenv~=1.0